requirements.txt
accelerate==1.10.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
diffusers==0.35.1
distro==1.9.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
googletrans==4.0.2
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
ImageHash==4.3.2
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.4.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
openai==1.100.2
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
playwright==1.55.0
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pyparsing==3.2.3
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytrends==4.9.2
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
replicate==1.0.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.6.2
schedule==1.2.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.48.1
streamlit-lottie==0.0.5
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
tweepy==4.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
win32_setctime==1.2.0
zipp==3.23.0
zstandard==0.24.0
